Q2/next/requirements.txt
djangorestframework
djoser
drf-yasg
pillow
